requirements_test.txt
pytest>=8.0
pytest-cov>=5.0
pytest-homeassistant-custom-component>=0.13.0
defusedxml==0.7.1
